wrap search export
